edit Auth

+++ src/main/java/org/sopt/server/api/UserController.java
@@ -1,9 +1,9 @@
 package org.sopt.server.api;
 
 import lombok.extern.slf4j.Slf4j;
+import org.sopt.server.dto.User;
 import org.sopt.server.model.DefaultRes;
 import org.sopt.server.model.SignUpReq;
-import org.sopt.server.model.test;
 import org.sopt.server.service.UserService;
 import org.sopt.server.utils.ResponseMessage;
 import org.sopt.server.utils.StatusCode;
@@ -13,10 +13,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.multipart.MultipartRequest;
 
 import javax.servlet.http.HttpServletRequest;
-import java.io.IOException;
 
 /**
  * Created by ds on 2018-10-23.
@@ -30,6 +28,7 @@ public class UserController {
     private static final String HEADER = "Authorization";
 
     private static final DefaultRes FAIL_DEFAULT_RES = new DefaultRes(StatusCode.INTERNAL_SERVER_ERROR, ResponseMessage.INTERNAL_SERVER_ERROR);
+    private static final DefaultRes UNAUTHORIZED_RES = new DefaultRes(StatusCode.UNAUTHORIZED, ResponseMessage.UNAUTHORIZED);
 
     private final UserService userService;
 
@@ -39,22 +38,20 @@ public class UserController {
 
     /**
      * 마이 페이지 조회
-     * @param httpServletRequest
-     * @param userIdx
-     * @return
+     *
+     * @param httpServletRequest Request
+     * @param userIdx 사용자 고유 번호
+     * @return ResponseEntity
      */
     @GetMapping("/{userIdx}")
-    public ResponseEntity getMyPage(
-            final HttpServletRequest httpServletRequest,
-            @PathVariable final int userIdx
-    ) {
+    public ResponseEntity getMyPage(final HttpServletRequest httpServletRequest, @PathVariable final int userIdx) {
         try {
-            final String jwt = httpServletRequest.getHeader(HEADER);
-            if (JwtUtils.checkAuth(jwt, userIdx).getData())
-                return new ResponseEntity<>(userService.findByUserIdx(JwtUtils.decode(jwt).get().getUser_idx(), userIdx), HttpStatus.OK);
-            return new ResponseEntity<>(userService.findByUserIdx(0, userIdx), HttpStatus.OK);
+            final int tokenValue = JwtUtils.decode(httpServletRequest.getHeader(HEADER)).getUser_idx();
+            DefaultRes<User> defaultRes = userService.findByUserIdx(userIdx);
+            if (tokenValue == userIdx) defaultRes.getData().setAuth(true);
+            return new ResponseEntity<>(defaultRes, HttpStatus.OK);
         } catch (Exception e) {
-            log.info(e.getMessage());
+            log.error(e.getMessage());
             return new ResponseEntity<>(FAIL_DEFAULT_RES, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
@@ -62,40 +59,41 @@ public class UserController {
     /**
      * 회원 가입
      *
-     * @param
-     * @return
+     * @param signUpReq 회원 정보
+     * @return ResponseEntity
      */
     @PostMapping("")
-    public ResponseEntity signUp(final SignUpReq signUpReq) throws IOException {
+    public ResponseEntity signUp(final SignUpReq signUpReq) {
         try {
             return new ResponseEntity<>(userService.save(signUpReq), HttpStatus.OK);
         } catch (Exception e) {
-            log.info(e.getMessage());
+            log.error(e.getMessage());
             return new ResponseEntity<>(FAIL_DEFAULT_RES, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     /**
-     * 내 정보 수정
+     * 회원 정보 수정
      *
-     * @param userIdx
-     * @return
+     * @param jwt       토큰
+     * @param userIdx   사용자 고유 번호
+     * @param signUpReq 회원 내용
+     * @param profile   사진
+     * @return ResponseEntity
      */
     @Auth
     @PutMapping("/{userIdx}")
-    public ResponseEntity updateMyPage(
-            @RequestHeader("Authorization") final String jwt,
-            @PathVariable final int userIdx,
-            final SignUpReq signUpReq,
-            @RequestPart(value = "profile", required = false) final MultipartFile profile) throws IOException {
+    public ResponseEntity updateMyPage(@RequestHeader("Authorization") final String jwt, @PathVariable final int userIdx,
+                                       final SignUpReq signUpReq,
+                                       @RequestPart(value = "profile", required = false) final MultipartFile profile) {
         try {
-            if(profile != null) signUpReq.setProfile(profile);
-            final DefaultRes<Boolean> defaultRes = JwtUtils.checkAuth(jwt, userIdx);
-            if (defaultRes.getData())
+            if (profile != null) signUpReq.setProfile(profile);
+            final int tokenValue = JwtUtils.decode(jwt).getUser_idx();
+            if (tokenValue == userIdx)
                 return new ResponseEntity<>(userService.update(userIdx, signUpReq), HttpStatus.OK);
-            return new ResponseEntity<>(defaultRes, HttpStatus.OK);
+            return new ResponseEntity<>(UNAUTHORIZED_RES, HttpStatus.OK);
         } catch (Exception e) {
-            log.info(e.getMessage());
+            log.error(e.getMessage());
             return new ResponseEntity<>(FAIL_DEFAULT_RES, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
@@ -103,21 +101,19 @@ public class UserController {
     /**
      * 회원 탈퇴
      *
-     * @param userIdx
-     * @return
+     * @param jwt     토큰
+     * @param userIdx 사용자 고유 번호
+     * @return ResponseEntity
      */
     @Auth
     @DeleteMapping("/{userIdx}")
-    public ResponseEntity goodBye(
-            @RequestHeader("Authorization") final String jwt,
-            @PathVariable final int userIdx) {
+    public ResponseEntity goodBye(@RequestHeader("Authorization") final String jwt, @PathVariable final int userIdx) {
         try {
-            final DefaultRes<Boolean> defaultRes = JwtUtils.checkAuth(jwt, userIdx);
-            if (defaultRes.getData())
-                return new ResponseEntity<>(userService.deleteByUserIdx(userIdx), HttpStatus.OK);
-            return new ResponseEntity<>(defaultRes, HttpStatus.OK);
+            final int tokenValue = JwtUtils.decode(jwt).getUser_idx();
+            if (tokenValue == userIdx) return new ResponseEntity<>(userService.deleteByUserIdx(userIdx), HttpStatus.OK);
+            return new ResponseEntity<>(UNAUTHORIZED_RES, HttpStatus.OK);
         } catch (Exception e) {
-            log.info(e.getMessage());
+            log.error(e.getMessage());
             return new ResponseEntity<>(FAIL_DEFAULT_RES, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
